% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tikz.R
\name{save_tikz}
\alias{save_tikz}
\title{Function for saving plots as tikz file.}
\usage{
save_tikz(
  plot,
  filename = "plot",
  dir = "written/figures",
  height,
  width,
  file_out = NULL,
  sanitize = F
)
}
\arguments{
\item{plot}{A plot.}

\item{filename}{The name of the file to be created.}

\item{dir}{The directory where the file should be saved, without trailing "/"}

\item{height}{The height of th plot in inches.}

\item{width}{Width of the plot in inches.}

\item{file_out}{Variable in order to track output in drake plan with file_out().}

\item{sanitize}{Clean out special LaTeX characters if they throw an error.}
}
\value{
A file in the specified folder.
}
\description{
Function for saving plots as tikz file.
}
