% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{setup_reticulate}
\alias{setup_reticulate}
\title{Sets up a python virtualenv adn Rprofile and configures reticulate
for use in the current directory/project}
\usage{
setup_reticulate()
}
\description{
Also see https://support.rstudio.com/hc/en-us/articles/360023654474-Installing-and-Configuring-Python-with-RStudio
}
