% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_latex.R
\name{use_latex}
\alias{use_latex}
\title{Creates files for latex integration in "written/".}
\usage{
use_latex(dir = "written")
}
\arguments{
\item{dir}{\code{string scaler}, giving the relative path to the directory
where the template should be created}
}
\description{
Creates files for latex integration in "written/".
}
