% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_project.R
\name{use_project}
\alias{use_project}
\title{Function for creating project folders for writing or data analysis.}
\usage{
use_project(path = "~/Documents/Projects/", project = "template",
  mode = "tex_only")
}
\arguments{
\item{path}{Path where the project files should be created. Defauts to "~/Documents/Projects/".}

\item{project}{Project name}

\item{mode}{"tex_only" only creates the files for writing a latex document. "data_minimal"
Creates and RProject with git, magrittr::pipe integration and a data-raw folder. "data_full"
adds drake and LaTeX integration.}
}
\description{
Function for creating project folders for writing or data analysis.
}
