% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rmarkdown_paper.R
\name{use_rmarkdown_paper}
\alias{use_rmarkdown_paper}
\title{Creates files for latex integration in "written/".}
\usage{
use_rmarkdown_paper(dir = "written")
}
\arguments{
\item{dir}{\code{string scaler}, giving the relative path to the directory
where the template should be created}
}
\description{
Creates files for latex integration in "written/".
}
