% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_latex.R
\name{use_latex}
\alias{use_latex}
\title{Create files for latex integration in "written/"}
\usage{
use_latex(path, project)
}
\arguments{
\item{path}{The Path to the project folder.}

\item{project}{The name of the project folder}
}
\description{
Create files for latex integration in "written/"
}
