% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_xaringan.R
\name{use_xaringan}
\alias{use_xaringan}
\title{Creates files for xaringan html presentation integration in "presentation/".}
\usage{
use_xaringan(dir = "presentation")
}
\description{
Creates files for xaringan html presentation integration in "presentation/".
}
